Move DirectoryPreparationStrategy
