+++ Cql/Cql.Runtime/IO/DirectoryInfoHandler.cs
@@ -1,0 +1,7 @@
+﻿namespace Hl7.Cql.Runtime.IO;
+
+/// <summary>
+/// Represents a method that handles a <see cref="DirectoryInfo"/> object.
+/// </summary>
+/// <param name="directory">The directory to handle.</param>
+public delegate void DirectoryInfoHandler(DirectoryInfo directory);
